Add KillMode=none in some services
This is workaround for DrKonqi, systemd assumes that service is dead when dbus address
is not visible on bus and tries to kill its child processes, but when something like
plasmashell crashes we still need its processes alive to get backtrace from DrKonqi.

This is workaround and should be replaced with proper solution in KCrash/processes.
Why this is workaround? systemd should be controlling processes but with this we trust
DrKonqi/kquitapp5 that they will do their cleanup

https://techbase.kde.org/User:Eliasp/Drafts/Plasma_Workspace_as_systemd_user-session#systemd.2FKCrash_race_conditions